requirements.txt
Flask==3.0.3
Requests==2.32.3
setuptools==65.5.0
python-dotenv==1.0.1
